+++ src/main/kotlin/chromia/RellVersionedLanguageServerFactory.kt
@@ -28,5 +28,4 @@ abstract class RellVersionedLanguageServerFactory(private val version: RellVersi
     override fun createClientFeatures(): LSPClientFeatures = RellLspClientFeatures()
 }
 
-class Rell0160LanguageServerFactory : RellVersionedLanguageServerFactory(RellVersion(0, 16, 0))
 class Rell0161LanguageServerFactory : RellVersionedLanguageServerFactory(RellVersion(0, 16, 1))